Added failing index tests to show that indexes can be created without configure indexes being called. (#143)
* Added failing index tests to show that indexes can be created without  configure indexes being called.

* Added more test cases

* Updated tests

* Logs index creation response.

Ensures the Elasticsearch index creation response is logged regardless of success or failure. This improves debugging and monitoring of index creation processes.

* Ensures index exists before patching document.

Ensures that the Elasticsearch index exists before attempting to patch a document. This prevents potential errors and improves the reliability of the patching operation.

* Handles ObjectId strings for date retrieval

Extends the index determination logic to handle string representations of ObjectIds.

This ensures that date-based indices can be correctly retrieved when the target is a string that can be parsed as an ObjectId.

* Enhances index existence checks in tests

Improves the index tests by adding more specific index existence assertions.
It now verifies the existence of both the index alias and the concrete index for specific document ids, where applicable.
This provides more accurate and reliable test results.

* Removes unnecessary trace logging

Removes default trace logging to reduce noise in test output.
Trace logging is now configured only for the EmployeeRepository.

* Improves handling of missing documents in patch operations.

Ensures that `DocumentNotFoundException` is thrown when attempting to patch a non-existent document in Elasticsearch.

Previously, a generic `DocumentException` was thrown, which didn't clearly indicate the root cause.

This change introduces a check for document existence before attempting to patch and throws a `DocumentNotFoundException` if the document is not found.

This provides more specific error handling and aligns with expected behavior.

Also, updates resilience policy to handle DocumentNotFoundException.

* Improves exception handling for patch operations

Adds a specific DocumentNotFoundException when a document isn't found during a patch operation, allowing for more targeted error handling.

This change enhances the robustness of patch operations by providing clearer feedback when a document does not exist.

* Fixes index creation during patch all by query.

Updates PatchAllAsync test to assert that daily indices are not created when using PatchAllAsync with a query across multiple days.
The test was incorrectly asserting that these indices would be created, which is not the expected behavior.

* Cleans up Elasticsearch alias handling

Removes unnecessary workaround for Elasticsearch-net issue.
The workaround is no longer needed and simplifies the alias retrieval logic.

Adds a list initialization for clarity.

* Adds logging for Elasticsearch requests

Adds logging for all Elasticsearch requests and responses.

This change ensures that all communication with Elasticsearch is logged, which is crucial for debugging, monitoring, and auditing purposes.
The log level can be configured using options for each request.

* Logs Elasticsearch request results

Ensures that Elasticsearch request results are logged, even when the requests are valid.
This improves debugging and monitoring capabilities.

* Updates xUnit and Jint packages

Updates the xunit.runner.visualstudio package to the latest version to improve test execution.
Also updates the Jint package to address potential issues and incorporate recent improvements.